# internal/docs/swagger.yaml
definitions:
...
  request.CasbinInReceive:
    properties:
      authorityId:
        description: 权限id
        type: integer
      casbinInfos:
        items:
          $ref: '#/definitions/request.CasbinInfo'
        type: array
    type: object
  request.CasbinInfo:
    properties:
      method:
        description: 方法
        type: string
      path:
        description: 路径
        type: string
    type: object
...
  request.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
...
  request.SearchApiParams:
    properties:
      ID:
        description: 主键ID
        type: integer
      apiGroup:
        description: api组
        type: string
      createdAt:
        description: 创建时间
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      description:
        description: api中文描述
        type: string
      keyword:
        description: 关键字
        type: string
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      path:
        description: api路径
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.PolicyPathResponse:
    properties:
      paths:
        items:
          $ref: '#/definitions/request.CasbinInfo'
        type: array
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.SysAuthorityResponse:
    properties:
      authority:
        $ref: '#/definitions/system.SysAuthority'
    type: object
...
  system.Meta:
    properties:
      activeName:
        type: string
      closeTab:
        description: 自动关闭tab
        type: boolean
      defaultMenu:
        description: 是否是基础路由（开发中）
        type: boolean
      icon:
        description: 菜单图标
        type: string
      keepAlive:
        description: 是否缓存
        type: boolean
      title:
        description: 菜单名
        type: string
    type: object
  system.SysApi:
    properties:
      ID:
        description: 主键ID
        type: integer
      apiGroup:
        description: api组
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: api中文描述
        type: string
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      deletedAt:
        type: string
      menus:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      parentId:
        description: 父角色ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenu:
    properties:
      ID:
        description: 主键ID
        type: integer
      authoritys:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      children:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      component:
        description: 对应前端文件路径
        type: string
      createdAt:
        description: 创建时间
        type: string
      hidden:
        description: 是否在列表隐藏
        type: boolean
      menuBtn:
        items:
          $ref: '#/definitions/system.SysBaseMenuBtn'
        type: array
      meta:
        allOf:
          - $ref: '#/definitions/system.Meta'
        description: 附加属性
      name:
        description: 路由name
        type: string
      parameters:
        items:
          $ref: '#/definitions/system.SysBaseMenuParameter'
        type: array
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由path
        type: string
      sort:
        description: 排序标记
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuBtn:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      desc:
        type: string
      name:
        type: string
      sysBaseMenuID:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuParameter:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      key:
        description: 地址栏携带参数的key
        type: string
      sysBaseMenuID:
        type: integer
      type:
        description: 地址栏携带参数为params还是query
        type: string
      updatedAt:
        description: 更新时间
        type: string
      value:
        description: 地址栏携带参数的值
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.7.9-beta
paths:
  /api/createApi:
    post:
      consumes:
        - application/json
      parameters:
        - description: api路径, api中文描述, api组, 方法
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysApi'
      produces:
        - application/json
      responses:
        "200":
          description: 创建基础api
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 创建基础api
      tags:
        - SysApi
  /api/deleteApi:
    post:
      consumes:
        - application/json
      parameters:
        - description: ID
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysApi'
      produces:
        - application/json
      responses:
        "200":
          description: 删除api
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 删除api
      tags:
        - SysApi
...
  /api/getApiList:
    post:
      consumes:
        - application/json
      parameters:
        - description: 分页获取API列表
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.SearchApiParams'
      produces:
        - application/json
      responses:
        "200":
          description: 分页获取API列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.PageResult'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 分页获取API列表
      tags:
        - SysApi
...
  /api/updateApi:
    post:
      consumes:
        - application/json
      parameters:
        - description: api路径, api中文描述, api组, 方法
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysApi'
      produces:
        - application/json
      responses:
        "200":
          description: 修改基础api
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 修改基础api
      tags:
        - SysApi
...
  /authority/createAuthority:
    post:
      consumes:
        - application/json
      parameters:
        - description: 权限id, 权限名, 父角色id
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysAuthority'
      produces:
        - application/json
      responses:
        "200":
          description: 创建角色,返回包括系统角色详情
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.SysAuthorityResponse'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 创建角色
      tags:
        - Authority
  /authority/deleteAuthority:
    post:
      consumes:
        - application/json
      parameters:
        - description: 删除角色
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysAuthority'
      produces:
        - application/json
      responses:
        "200":
          description: 删除角色
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 删除角色
      tags:
        - Authority
  /authority/getAuthorityList:
    post:
      consumes:
        - application/json
      parameters:
        - description: 页码, 每页大小
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.PageInfo'
      produces:
        - application/json
      responses:
        "200":
          description: 分页获取角色列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.PageResult'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 分页获取角色列表
      tags:
        - Authority
  /authority/setDataAuthority:
    post:
      consumes:
        - application/json
      parameters:
        - description: 设置角色资源权限
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysAuthority'
      produces:
        - application/json
      responses:
        "200":
          description: 设置角色资源权限
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 设置角色资源权限
      tags:
        - Authority
  /authority/updateAuthority:
    put:
      consumes:
        - application/json
      parameters:
        - description: 权限id, 权限名, 父角色id
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/system.SysAuthority'
      produces:
        - application/json
      responses:
        "200":
          description: 更新角色信息,返回包括系统角色详情
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.SysAuthorityResponse'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 更新角色信息
      tags:
        - Authority
...
  /casbin/UpdateCasbin:
    post:
      consumes:
        - application/json
      parameters:
        - description: 权限id, 权限模型列表
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.CasbinInReceive'
      produces:
        - application/json
      responses:
        "200":
          description: 更新角色api权限
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 更新角色api权限
      tags:
        - Casbin
  /casbin/getPolicyPathByAuthorityId:
    post:
      consumes:
        - application/json
      parameters:
        - description: 权限id, 权限模型列表
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.CasbinInReceive'
      produces:
        - application/json
      responses:
        "200":
          description: 获取权限列表,返回包括casbin详情列表
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.PolicyPathResponse'
                  msg:
                    type: string
                type: object
      security:
        - ApiKeyAuth: []
      summary: 获取权限列表
      tags:
        - Casbin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...